+++ mobile/src/app/model/xphistory.ts
@@ -5,14 +5,14 @@ export type EventClass = 'A' | 'B' | 'C' | 'D';
 
 /** Description of the referee allocation job */
 export type RefereeAllocation =
-    /** The coach does not allocate the referees on game */
+    /** The coach does not allocate the referees on games */
     'No'
-    /** The coach contribute to the rreferee allocation */
+    /** The coach contribute to the referee allocation */
     | 'Contribute'
     /** The coach fully participate to the refeee allocation */
     | 'Yes';
 
-export interface CoachingHistory extends PersistentData {
+export interface Xp extends PersistentData {
     /** Identifier of the coach */
     coachId: string;
     /** Name of the event */
@@ -26,7 +26,8 @@ export interface CoachingDay {
     coachingDate: Date;
     /** Duration of the game in minutes */
     gameDuration: number;
-    /** Number of minute of coached game during the event */
+    /** Number of minute of coached game during the day */
     coachingDuration: number;
+    /** Indicate the contribution level of the referee coach to the referee allocation */
     refereeAllocation: RefereeAllocation;
 }