+++ release-notes.md
@@ -4,6 +4,7 @@ This file lists the development task to do:
 
 ## Possibles features
 
+- Assessment of coach (coach profile)
 - Mobile app installation on IOS device
 - Settings: limit to referee to a region or a group => defintion of referee group
 - Create user group used to simplify the sharing